03.eProjects/Example-Python-Django-V1/bikestore/requirements-copy.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
